bugged proj3

+++ src/main/java/io/github/timliiang/scene/Material.java
@@ -1,12 +1,12 @@
 package io.github.timliiang.scene;
 
 public class Material {
-    private Color ambient; // ambient constant, percentage of ambient light the material reflects
-    private Color diffuse; // diffuse constant, percentage of diffuse light the material reflects
-    private Color specular; // specular constant, percentage of specular light the material reflects
-    private float shininess; // shininess factor, higher value means shinier material
+    private Color ambient; // k_a, ambient constant, percentage of ambient light the material reflects
+    private Color diffuse; // k_d, diffuse constant, percentage of diffuse light the material reflects
+    private Color specular; // k_s, specular constant, percentage of specular light the material reflects
+    private int shininess; // alpha, shininess factor, higher value means shinier material
 
-    public Material(Color ambient, Color diffuse, Color specular, float shininess) {
+    public Material(Color ambient, Color diffuse, Color specular, int shininess) {
         this.ambient = ambient;
         this.diffuse = diffuse;
         this.specular = specular;
@@ -37,11 +37,11 @@ public class Material {
         this.specular = specular;
     }
 
-    public float getShininess() {
+    public int getShininess() {
         return this.shininess;
     }
 
-    public void setShininess(float shininess) {
+    public void setShininess(int shininess) {
         this.shininess = shininess;
     }
